requirements.txt
# InsurMinds Saúde — Dependências Python
# Instale com: pip install -r requirements.txt

python-telegram-bot==21.5
langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface==0.0.3
huggingface-hub==0.24.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
langchain-google-genai
flask
